web_apps/python/requirements.txt
fastapi==0.85.0
strawberry-graphql==0.133.1
uvicorn[standard]==0.18.3
python-multipart==0.0.6